requirements.txt
git+ssh://git@github.com/orlevi111/mest@reinvnet#egg=mest[falcon]
minio==5.0.*
python-dotenv==0.13.*
